Remove DetourCrowd module and migrate RecastDemo to SDL3
- Delete DetourCrowd/ source tree, Tests/DetourCrowd/, and Tool_Crowd
- Remove all DetourCrowd references from CMakeLists.txt, premake5.lua,
  recastnavigation.pc.in, Doxyfile, README, and docs
- Migrate RecastDemo from SDL2 to SDL3: update window/event/GL APIs,
  swap imgui_impl_sdl2 for imgui_impl_sdl3 (added from imgui v1.92.0),
  update Xcode project to reference local SDL3.framework
- Inline dtMergeCorridorStartMoved into Tool_NavMeshTester.cpp (was
  provided by the now-deleted DetourPathCorridor.h)

+++ RecastDemo/Include/AppState.h
@@ -33,7 +33,7 @@ struct AppState
 
 	// Time
 	float timeAcc = 0.0f;
-	Uint32 prevFrameTime = 0;
+	Uint64 prevFrameTime = 0;
 
 	// Input
 	int mousePos[2]{0, 0};
